+++ crates/config/src/config.rs
@@ -1,11 +1,13 @@
 //! Configuration files.
 
+use eyre::eyre;
 use reth_network_types::{PeersConfig, SessionsConfig};
 use reth_prune_types::PruneModes;
 use reth_stages_types::ExecutionStageThresholds;
 use serde::{Deserialize, Deserializer, Serialize};
 use std::{
     ffi::OsStr,
+    fs,
     path::{Path, PathBuf},
     time::Duration,
 };
@@ -29,6 +31,31 @@ pub struct Config {
 }
 
 impl Config {
+    /// Load a [`Config`] from a specified path.
+    ///
+    /// A new configuration file is created with default values if none
+    /// exists.
+    pub fn from_path(path: impl AsRef<Path>) -> eyre::Result<Self> {
+        let path = path.as_ref();
+        match fs::read_to_string(path) {
+            Ok(cfg_string) => {
+                toml::from_str(&cfg_string).map_err(|e| eyre!("Failed to parse TOML: {e}"))
+            }
+            Err(e) if e.kind() == std::io::ErrorKind::NotFound => {
+                if let Some(parent) = path.parent() {
+                    fs::create_dir_all(parent)
+                        .map_err(|e| eyre!("Failed to create directory: {e}"))?;
+                }
+                let cfg = Self::default();
+                let s = toml::to_string_pretty(&cfg)
+                    .map_err(|e| eyre!("Failed to serialize to TOML: {e}"))?;
+                fs::write(path, s).map_err(|e| eyre!("Failed to write configuration file: {e}"))?;
+                Ok(cfg)
+            }
+            Err(e) => Err(eyre!("Failed to load configuration: {e}")),
+        }
+    }
+
     /// Returns the [`PeersConfig`] for the node.
     ///
     /// If a peers file is provided, the basic nodes from the file are added to the configuration.
@@ -390,7 +417,7 @@ mod tests {
     use super::{Config, EXTENSION};
     use reth_network_peers::TrustedPeer;
     use reth_node_core::node_config::NodeConfig;
-    use std::{str::FromStr, time::Duration};
+    use std::{path::Path, str::FromStr, time::Duration};
 
     fn with_tempdir(filename: &str, proc: fn(&std::path::Path)) {
         let temp_dir = tempfile::tempdir().unwrap();
@@ -399,6 +426,82 @@ mod tests {
         proc(&config_path);
 
         temp_dir.close().unwrap()
+    }
+
+    /// Run a test function with a temporary config path as fixture.
+    fn with_config_path(test_fn: fn(&Path)) {
+        // Create a temporary directory for the config file
+        let config_dir = tempfile::tempdir().expect("creating test fixture failed");
+        // Create the config file path
+        let config_path =
+            config_dir.path().join("example-app").join("example-config").with_extension("toml");
+        // Run the test function with the config path
+        test_fn(&config_path);
+        config_dir.close().expect("removing test fixture failed");
+    }
+
+    #[test]
+    fn test_load_path_works() {
+        with_config_path(|path| {
+            let config = Config::from_path(path).expect("load_path failed");
+            assert_eq!(config, Config::default());
+        })
+    }
+
+    #[test]
+    fn test_load_path_reads_existing_config() {
+        with_config_path(|path| {
+            let config = Config::default();
+
+            // Create the parent directory if it doesn't exist
+            if let Some(parent) = path.parent() {
+                std::fs::create_dir_all(parent).expect("Failed to create directories");
+            }
+
+            // Write the config to the file
+            std::fs::write(path, toml::to_string(&config).unwrap())
+                .expect("Failed to write config");
+
+            // Load the config from the file and compare it
+            let loaded = Config::from_path(path).expect("load_path failed");
+            assert_eq!(config, loaded);
+        })
+    }
+
+    #[test]
+    fn test_load_path_fails_on_invalid_toml() {
+        with_config_path(|path| {
+            let invalid_toml = "invalid toml data";
+
+            // Create the parent directory if it doesn't exist
+            if let Some(parent) = path.parent() {
+                std::fs::create_dir_all(parent).expect("Failed to create directories");
+            }
+
+            // Write invalid TOML data to the file
+            std::fs::write(path, invalid_toml).expect("Failed to write invalid TOML");
+
+            // Attempt to load the config should fail
+            let result = Config::from_path(path);
+            assert!(result.is_err());
+        })
+    }
+
+    #[test]
+    fn test_load_path_creates_directory_if_not_exists() {
+        with_config_path(|path| {
+            // Ensure the directory does not exist
+            let parent = path.parent().unwrap();
+            assert!(!parent.exists());
+
+            // Load the configuration, which should create the directory and a default config file
+            let config = Config::from_path(path).expect("load_path failed");
+            assert_eq!(config, Config::default());
+
+            // The directory and file should now exist
+            assert!(parent.exists());
+            assert!(path.exists());
+        });
     }
 
     #[test]
@@ -434,7 +537,7 @@ mod tests {
             .expect("Failed to write config file");
 
             // Load the config from the file
-            let loaded_config: Config = NodeConfig::load_path(config_path).unwrap();
+            let loaded_config = Config::from_path(config_path).unwrap();
 
             // Compare the loaded config with the original config
             assert_eq!(config, loaded_config);
@@ -455,7 +558,7 @@ mod tests {
             .expect("Failed to write config file");
 
             // Load the config from the file
-            let loaded_config: Config = NodeConfig::load_path(config_path).unwrap();
+            let loaded_config = Config::from_path(config_path).unwrap();
 
             // Compare the loaded config with the original config
             assert_eq!(config, loaded_config);
