+++ crates/node/core/src/node_config.rs
@@ -416,91 +416,3 @@ impl Default for NodeConfig {
         }
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use serde::Deserialize;
-
-    #[derive(PartialEq, Default, Debug, Serialize, Deserialize)]
-    struct ExampleConfig {
-        name: String,
-        count: usize,
-    }
-
-    /// Run a test function with a temporary config path as fixture.
-    fn with_config_path(test_fn: fn(&Path)) {
-        // Create a temporary directory for the config file
-        let config_dir = tempfile::tempdir().expect("creating test fixture failed");
-        // Create the config file path
-        let config_path =
-            config_dir.path().join("example-app").join("example-config").with_extension("toml");
-        // Run the test function with the config path
-        test_fn(&config_path);
-        config_dir.close().expect("removing test fixture failed");
-    }
-
-    #[test]
-    fn test_load_path_works() {
-        with_config_path(|path| {
-            let config: ExampleConfig = NodeConfig::load_path(path).expect("load_path failed");
-            assert_eq!(config, ExampleConfig::default());
-        })
-    }
-
-    #[test]
-    fn test_load_path_reads_existing_config() {
-        with_config_path(|path| {
-            let config: ExampleConfig = ExampleConfig { name: "Test".to_string(), count: 42 };
-
-            // Create the parent directory if it doesn't exist
-            if let Some(parent) = path.parent() {
-                std::fs::create_dir_all(parent).expect("Failed to create directories");
-            }
-
-            // Write the config to the file
-            std::fs::write(path, toml::to_string(&config).unwrap())
-                .expect("Failed to write config");
-
-            // Load the config from the file and compare it
-            let loaded = NodeConfig::load_path(path).expect("load_path failed");
-            assert_eq!(config, loaded);
-        })
-    }
-
-    #[test]
-    fn test_load_path_fails_on_invalid_toml() {
-        with_config_path(|path| {
-            let invalid_toml = "invalid toml data";
-
-            // Create the parent directory if it doesn't exist
-            if let Some(parent) = path.parent() {
-                std::fs::create_dir_all(parent).expect("Failed to create directories");
-            }
-
-            // Write invalid TOML data to the file
-            std::fs::write(path, invalid_toml).expect("Failed to write invalid TOML");
-
-            // Attempt to load the config should fail
-            let result: eyre::Result<ExampleConfig> = NodeConfig::load_path(path);
-            assert!(result.is_err());
-        })
-    }
-
-    #[test]
-    fn test_load_path_creates_directory_if_not_exists() {
-        with_config_path(|path| {
-            // Ensure the directory does not exist
-            let parent = path.parent().unwrap();
-            assert!(!parent.exists());
-
-            // Load the configuration, which should create the directory and a default config file
-            let config: ExampleConfig = NodeConfig::load_path(path).expect("load_path failed");
-            assert_eq!(config, ExampleConfig::default());
-
-            // The directory and file should now exist
-            assert!(parent.exists());
-            assert!(path.exists());
-        });
-    }
-}
